DVISA-597: (feature) rename
