+++ src/tools/annotation/AngleTool.js
@@ -274,7 +274,7 @@ export default class AngleTool extends BaseAnnotationTool {
       const str = '00B0'; // Degrees symbol
 
       return (
-        toGermanStringTemp(data.rAngle) +
+        toGermanNumberStringTemp(data.rAngle) +
         String.fromCharCode(parseInt(str, 16)) +
         suffix
         //data.rAngle.toString() + String.fromCharCode(parseInt(str, 16)) + suffix
