+++ src/tools/annotation/CobbAngleTool.js
@@ -28,7 +28,7 @@ import triggerEvent from '../../util/triggerEvent.js';
 import throttle from '../../util/throttle';
 import getPixelSpacing from '../../util/getPixelSpacing';
 import { getModule } from '../../store/index';
-import toGermanStringTemp from '../../util/toGermanStringTemp.js';
+import toGermanNumberStringTemp from '../../util/toGermanNumberStringTemp.js';
 
 /**
  * @public
@@ -387,7 +387,7 @@ export default class CobbAngleTool extends BaseAnnotationTool {
       const str = '00B0'; // Degrees symbol
 
       return (
-        toGermanStringTemp(rAngle) /*rAngle.toString()*/ +
+        toGermanNumberStringTemp(rAngle) /*rAngle.toString()*/ +
         String.fromCharCode(parseInt(str, 16)) +
         suffix
       );
