+++ src/tools/annotation/FreehandRoiTool.js
@@ -29,6 +29,7 @@ import freehandUtils from '../../util/freehand/index.js';
 import { getLogger } from '../../util/logger.js';
 import throttle from '../../util/throttle';
 import { getModule } from '../../store/index';
+import toGermanNumberStringTemp from '../../util/toGermanNumberStringTemp.js';
 
 const logger = getLogger('tools:annotation:FreehandRoiTool');
 
@@ -520,11 +521,11 @@ export default class FreehandRoiTool extends BaseAnnotationTool {
 
         // Create a line of text to display the mean and any units that were specified (i.e. HU)
 
-        let meanText = `avg: ${toGermanStringTemp(
+        let meanText = `avg: ${toGermanNumberStringTemp(
           meanStdDev.mean
         )} ${moSuffix}`; //`Mean: ${numbersWithCommas(meanStdDev.mean.toFixed(2))} ${moSuffix}`;
         // Create a line of text to display the standard deviation and any units that were specified (i.e. HU)
-        let stdDevText = `sd: ${toGermanStringTemp(
+        let stdDevText = `sd: ${toGermanNumberStringTemp(
           meanStdDev.stdDev
         )} ${moSuffix}`; //`StdDev: ${numbersWithCommas(meanStdDev.stdDev.toFixed(2))} ${moSuffix}`;
 
@@ -555,7 +556,7 @@ export default class FreehandRoiTool extends BaseAnnotationTool {
         }
 
         // Create a line of text to display the area and its units
-        const areaText = `A: ${toGermanStringTemp(area)} ${suffix}`; //`Area: ${numbersWithCommas(area.toFixed(2))}${suffix}`;
+        const areaText = `A: ${toGermanNumberStringTemp(area)} ${suffix}`; //`Area: ${numbersWithCommas(area.toFixed(2))}${suffix}`;
 
         // Add this text line to the array to be displayed in the textbox
         textLines.push(areaText);
